Fixed data types of templated metrics.

+++ src/LinqAn.Google/Metrics/ContentGroupUniqueViews3.cs
@@ -6,7 +6,7 @@ namespace LinqAn.Google.Metrics
 	/// 	The number of unique content group views. Content group views within different sessions count as separate unique content view. This takes into account both the pagePath and pageTitle to determine uniqueness.
 	/// </summary>
 	[Description("The number of unique content group views. Content group views within different sessions count as separate unique content view. This takes into account both the pagePath and pageTitle to determine uniqueness.")]
-	public class ContentGroupUniqueViews3: Metric<float>
+	public class ContentGroupUniqueViews3: Metric<int>
 	{
 		/// <summary>
 		/// 	Instantiates a <seealso cref="ContentGroupUniqueViews3" />.
